Reproduction for the bug
The page renders a component 3 times but using different delays in their Resource to cause an out of order error. Here the CSS for the list and item components are included next to the third rendered list because it is the once that resolves first but then the HTML that is streamed to the client will show the first two lists as unstyled content until the stream reaches the last list.

Signed-off-by: Jeremy Wickersheimer <jwickers@gmail.com>